(PC-20917)[PRO] chore: merge types util file from adage to pro
