+++ pro/src/utils/types.ts
@@ -1,3 +1,18 @@
+export interface ResultType {
+  objectID: string
+  offer: {
+    dates: number[]
+    name: string
+    thumbUrl: string
+  }
+  venue: {
+    name: string
+    publicName: string
+  }
+  isTemplate: boolean
+  __queryID: string
+}
+
 export const hasProperty = <T extends string>(
   element: unknown,
   property: T
